v2.0.0

+++ src/index.js
@@ -6,12 +6,9 @@ const _ = require("lodash"),
   path = require("path");
 
 const collectFunctionEnvVariables = require("./lib/collectFunctionEnvVariables");
-const setEnvVariables = require("./lib/setEnvVariables");
-const collectOfflineEnvVariables = require("./lib/collectOfflineEnvVariables");
 const resolveCloudFormationEnvVariables = require("./lib/resolveCloudFormationEnvVariables");
+const setEnvVariables = require("./lib/setEnvVariables");
 const transformEnvVarsToString = require("./lib/transformEnvVarsToString");
-const collectResourcesOutputs = require("./lib/collectResourcesOutputs");
-const collectStackOutputs = require("./lib/collectStackOutputs");
 
 /**
  * Serverless Plugin to extract Serverless' Lambda environment variables into
@@ -41,7 +38,22 @@ class ExportEnv {
     };
 
     this.environmentVariables = {};
-    this.envFileName = ".env";
+    this.filename = ".env";
+    this.enableOffline = true;
+    this.overwriteExisting = true;
+    this.refMap = {};
+    this.getAttMap = {};
+    this.importValueMap = {};
+  }
+
+  loadConfig() {
+    const params = _.get(this.serverless, "service.custom.export-env");
+    this.filename = _.get(params, "filename", this.filename);
+    this.enableOffline = _.get(params, "enableOffline", this.enableOffline);
+    this.overwriteExisting = _.get(params, "overwriteExisting", this.overwriteExisting);
+    this.refMap = _.get(params, "refMap", this.refMap);
+    this.getAttMap = _.get(params, "getAttMap", this.getAttMap);
+    this.importValueMap = _.get(params, "importValueMap", this.importValueMap);
   }
 
   initOfflineHook() {
@@ -56,43 +68,36 @@ class ExportEnv {
     return BbPromise.try(() => {
       const envVars = {};
 
-      return collectStackOutputs(this.serverless).then((stackOutputs) => {
-        // collect global environment variables
-        const globalEnvironment = this.serverless.service.provider.environment;
-        _.assign(envVars, globalEnvironment);
+      // collect global environment variables
+      const globalEnvironment = this.serverless.service.provider.environment;
+      _.assign(envVars, globalEnvironment);
 
-        // collect Resources Outputs
-        const resourcesOutputs = collectResourcesOutputs(this.serverless, stackOutputs);
-        _.assign(envVars, resourcesOutputs);
+      // collect environment variables of functions
+      const functionEnvironment = collectFunctionEnvVariables(this.serverless);
+      _.assign(envVars, functionEnvironment);
 
-        // collect environment variables of functions
-        const functionEnvironment = collectFunctionEnvVariables(this.serverless);
-        _.assign(envVars, functionEnvironment);
-
-        // collect environment variables for serverless offline
-        if (this.isOfflineHooked) {
-          const offlineEnvVars = collectOfflineEnvVariables(this.serverless, this.options);
-          _.assign(envVars, offlineEnvVars);
-        }
-
-        process.env.SLS_DEBUG && this.serverless.cli.log(`Found ${_.size(envVars)} environment variable(s)`);
-        this.environmentVariables = envVars;
-        return BbPromise.resolve();
-      });
+      process.env.SLS_DEBUG && this.serverless.cli.log(`Found ${_.size(envVars)} environment variable(s)`);
+      this.environmentVariables = envVars;
+      return BbPromise.resolve();
     });
   }
 
   resolveEnvVars() {
-    // resolve environment variables referencing CloudFormation
-    return resolveCloudFormationEnvVariables(this.serverless, this.environmentVariables)
+    return resolveCloudFormationEnvVariables(this.serverless, this.environmentVariables, {
+      refMap: this.refMap,
+      getAttMap: this.getAttMap,
+      importValueMap: this.importValueMap,
+    })
       .then((resolved) => (this.environmentVariables = resolved))
       .return();
   }
 
   applyEnvVars() {
     return BbPromise.try(() => {
+      this.loadConfig();
+
       // If this is a local lambda invoke, replace the service environment with the resolved one
-      if (this.isOfflineHooked) {
+      if (this.isOfflineHooked && this.enableOffline) {
         setEnvVariables(this.serverless, this.environmentVariables);
       }
     });
@@ -100,23 +105,19 @@ class ExportEnv {
 
   writeEnvVars() {
     return BbPromise.try(() => {
-      process.env.SLS_DEBUG && this.serverless.cli.log("Writing .env file");
+      this.loadConfig();
 
-      const params = _.get(this.serverless, 'service.custom.export-env');
+      const envFilePath = path.resolve(this.serverless.config.servicePath, this.filename);
+      if (this.isOfflineHooked) {
+        // not writing anything by default
+      } else if (!fs.existsSync(envFilePath) || this.overwriteExisting) {
+        process.env.SLS_DEBUG && this.serverless.cli.log(`Writing ${this.filename} file`);
 
-      let filename = this.envFileName;
-      let pathFromRoot = "";
-
-      if (params != null) {
-        if (params.filename != null) filename = params.filename;
-        if (params.pathFromRoot != null) pathFromRoot = params.pathFromRoot;
+        const envDocument = transformEnvVarsToString(this.environmentVariables);
+        fs.writeFileSync(envFilePath, envDocument);
+      } else {
+        process.env.SLS_DEBUG && this.serverless.cli.log(`${this.filename} already exists. Leaving it untouched.`);
       }
-
-      const envFilePath = path.resolve(this.serverless.config.servicePath, pathFromRoot, filename);
-
-      const envDocument = transformEnvVarsToString(this.environmentVariables);
-
-      fs.writeFileSync(envFilePath, envDocument);
     });
   }
 }
